requirements.txt
Flask
cryptography
pycryptodome
SQLAlchemy